+++ src/App.css
@@ -1,5 +1,78 @@
-.GroupCom{
-    background-color: #f7ecfe;
-    margin: 15px;
-}
+/* App.css */
 
+/* Remove browser defaults */
+* {
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
+  }
+  /* Style App.js wrapper */
+  .App {
+    width: 100vw;
+    height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    overflow: hidden;
+  }
+  
+  /* Tab Container */
+  .Tabs {
+    height: auto;
+    min-height: 400px;
+    background: #FEF7FF;
+    margin: 3.5rem auto 1.5rem;
+    padding: 2rem 1rem;
+    color: #E8F0F2;
+    border-radius: 2rem;
+    @media (max-width: 769px) {
+      padding: 2rem 0;
+    }
+  
+  }
+  
+  /*Tab Navigation */
+  ul.nav {
+    width: 60%;
+    margin: 0 auto 2rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border: 1px solid #39A2DB;
+    border-radius: 2rem;
+    @media (max-width: 768px) {
+      width: 90%;
+    }
+  
+  
+  }
+  
+  
+  ul.nav li {
+    width: 50%;
+    padding: 1rem;
+    list-style: none;
+    text-align: left;
+    cursor: pointer;
+    transition: all 0.7s;
+    border-bottom-left-radius: 2rem;
+    border-top-left-radius: 2rem;
+  }
+  ul.nav li:nth-child(2) {
+    border-radius: 0;
+    border-bottom-right-radius: 2rem;
+    border-top-right-radius: 2rem;
+  }
+  ul.nav li:hover {
+    background: rgba(50, 224, 196, 0.15);
+  }
+  ul.nav li.active {
+    background: #39A2DB;
+  }
+  
+  /* First and Second Tab Styles */
+  .FirstTab p,
+  .SecondTab p {
+    font-size: 2rem;
+    text-align: left;
+  }
