+++ src/pages/Comentarios/ComentarioPage.css
@@ -1,0 +1,4 @@
+.GroupCom{
+    background-color: #f7ecfe;
+    margin: 15px;
+}
